Rename submodules under `cupy.linalg` package
