+++ cupy/linalg/_norms.py
