+++ cupy/linalg/_product.py
