+++ cupy/linalg/_util.py
